requirements.txt
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
decorator==4.4.1
emoji==0.5.4
ephem==3.7.7.1
future==0.18.2
pycparser==2.19
PySocks==1.7.1
python-telegram-bot==12.4.2
six==1.14.0
tornado==6.0.3
clarifai==2.6.2
